Moved all logic to its own command
With this change we are able to run the workflow commands
asynchronously, speeding up the process of saving the initial document.

This change does have the side effect that the tasks will be created a
bit later then the main document, the tasks may not be visible directly
but after a few seconds.